+++ base/strings/search.jl
@@ -448,16 +448,28 @@ julia> findprev("Julia", "JuliaLang", 6)
 findprev(t::AbstractString, s::AbstractString, i::Integer) = _rsearch(s, t, i)
 
 """
-    contains(haystack::AbstractString, needle::Union{AbstractString,Char})
+    contains(haystack::AbstractString, needle::Union{AbstractString,Regex,Char})
 
-Determine whether the second argument is a substring of the first.
+Determine whether the second argument is a substring of the first. If `needle`
+is a regular expression, checks whether `haystack` contains a match.
 
 # Examples
 ```jldoctest
 julia> contains("JuliaLang is pretty cool!", "Julia")
 true
+
+julia> contains("JuliaLang is pretty cool!", 'a')
+true
+
+julia> contains("aba", r"a.a")
+true
+
+julia> contains("abba", r"a.a")
+false
 ```
 """
+function contains end
+
 contains(haystack::AbstractString, needle::Union{AbstractString,Char}) =
     _searchindex(haystack, needle, start(haystack)) != 0